[sdk-tools] Use $HOME instead of mygetpwuid_r for default home dir
The comments for checkSSHConfig mention ${HOME} but the code
actually ignored this environment variable and went straight to
getting the actual home directory from /etc/passwd via mygetpwuid_r,
which is not always the same as ${HOME}. Now if the user has
changed ${HOME} the tools will use this correctly, and if not
available will fallback to the previous implementation.

This allows automated test tools to override ${HOME} and then
force all the output into a temporary directory instead of
reading/writing into the user home directory which may have
unintended side effects.

This implementation is now consistent with that used in
//tools/debug/symbolize/cmd/main.go which also writes to
the ${HOME}/.fuchsia directory.

Note that while fssh supports --data-path it only overrides the
location of the generated sshconfig file, but not the IdentityFile
location path which is still using the mygetpwuid_r home directory
and not $HOME as expected.

Bug: 111916
Change-Id: I06f4d12dc90762c021fc19b7ff719e8a635d1128
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748222
Commit-Queue: Wayne Piekarski <waynepie@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>