[kernel][vm] Mark accessed clean page about to be dirtied
When pages to be dirtied are accumulated in PrepareForWriteLocked, mark
any clean pages encountered as accessed, moving them to the front of the
page queue. This will protect them from eviction until the user pager
has a chance to respond to the DIRTY request, and potentially saves
redundant work from having to page the evicted page back in.

This only grants protection from eviction in the typical case where the
user pager is processing DIRTY requests at a reasonable rate. If there
are large delays such that the accessed pages age and qualify for
eviction again, they will be evicted i.e. this does not "pin" clean
pages.

Bug: 63989
Run-All-Tests: true
Change-Id: Ib67231523803ada13f53f4204611334d2cd8f96e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748967
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>