[nxpfmac] Handle deferred handling event
Mlan can send a defer handling event which means that we need to trigger
the main process when it arrives. This happens for things that are
initiated by our driver where mlan needs additional main processing to
happen for the operation to complete. Talking to firmware for example.

Handle this event and call on the bus to trigger main processing when it
happens.

Test: Included unit test
Change-Id: I45581f2e3d6c7fd6341ae07d56cce3b5b173088d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747612
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>