[vpython][yapf] Remove -vpython-interpreter from vpython invocation.
In a decision made today regarding the future of vpython in-tree, we've
decided to, in the short term, allow vpython's python version to diverge
from Fuchsia's vendored python.

The reason this is acceptable with respect to RFC-129 is that, in the
latest vpython, it uses a bundled python executable rather than
traversing $PATH for one, so it doesn't cause any nondeterminism in the
python version.

In a change to be merged shortly after this one, we'll roll the latest
version of vpython in-tree.

Bug: 107761
Change-Id: I39b91492672e8b151ac0ac6b66a74edda01c4217
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748102
Fuchsia-Auto-Submit: Catherine Duncan <catduncan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>