[nxpfmac] Fix IoctlAdapter timeout edge case
There was an edge case where a timeout could schedule a cancelation at
the same time as a completion was in progress. This lead to a double
call of IoctlComplete which in turn attempted to use the callback twice.
The code already protects from this by taking the callback pointer out
of the request, it can't be called twice. However the second attempt
would crash.

Fix this issue by checking if the callback is still available, if not
the ioctl complete turns into a noop and the actual completion wins.

Test: Included unit test
Change-Id: I8808fcc79847c867daf1b13f9639c830e8f57f79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747613
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>