[suspend_test] Use lid to trigger resume.
The lid driver sets itself as a wake device in its implementation of the
suspend hook, so using the lid to trigger resume tests that hook as well
as the lid as a wake device. We use the lid instead of the power button
for this test because using the lid to suspend and resume will likely be
the most common workflow for using suspend.

Also add more traces during suspend and resume so the logs are more
helpful when the test fails.

Bug: 107797

Change-Id: Ia6ee21da4e83a422e96b39445f7b60ef70d19fe1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748444
Reviewed-by: Natalie Eleftheriades <natlf@google.com>
Commit-Queue: Sarah Pham <smpham@google.com>