[fxfs] Make sure we zero the tail of blocks on reads
We need to ensure that the tail of blocks is always zeroed on reads.  We
also need to ensure that the tail of blocks is zeroed when extending a
file.

The code to zero the tail on writes has been removed on the basis that
it won't be adding *much* benefit since Fxfs and the underlying storage
stack is COW, so the data will likely be present somewhere anyway.

Fixed: 88676
Change-Id: Ie8e403dd7ade7bdf2c369f6d4d7389315cad7969
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744827
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Chris Suter <csuter@google.com>