[build][bazel] Record workspace events log.
This CL changes the generated `bazel` launcher script and .bazelrc
file to ensure that a log of workspace rule events is recorded each
time it is invoked.

This file is a binary that must be parsed with the Bazel Workspace
Log Parser program [1] to generate a text representation, that can
be used to detect undesirable download() and download_extract()
events triggered by workspace rules.

[1] https://github.com/bazelbuild/bazel/blob/master/src/tools/workspacelog/README.md

Bug: 113230
Change-Id: Ib92ddd585e074fa4c4842131c4f97d42681d5323
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/746123
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Turner <digit@google.com>