[component] Introduce `component_incoming` lib
This CL copies over the contents of `//sdk/lib/sys/component/cpp`
that pertains to FIDL client code into its own separate library.
The goal is to ultimately split up the library into an `incoming`
portion and an `outgoing` one. The motivation for this split is
that the `component::OutgoingDirectory` has a runtime dependency
on the `libsvc.so`. This is problematic for two reasons:

1: The shared library `libsvc.so` isn't always available to consumers
of the library. See the attached bug for a build breakage when trying
to build `libsyslog.so` which had a runtime dep on `libsvc.so`.

2: Not all components need the `OutgoingDirectory` portion of the
library, and its unnecessary to package a shared library that won't
be used.

Subsequent CL(s) will migrate all current users of `//sdk/lib/sys/component/cpp`
to this library.

Bug: 113114
Change-Id: I35f54939b40e6ce9f4931d648b610ad9c52e5e8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749342
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>