[bt][a2dp] Mark tracing as optional
This is a reland of commit 47341d71947f1a9fccc2732eb28639a99cc5e8d9
with the following changes:
 - Use tracing in source instead of sink in loopback as e2e test detects
   the packet sent tracing to verify data is sent.
 - Added comments to bt-a2dp-loopback to indicate this need.

fuchsia.tracing.provider.Registry is not available on all builds, and we
can continue to operate without it, so mark it as an optional protocol
in case tracing is unavailable.

Reformat optional protocols for brevity.

Bug: b/248584553
Test: updated bt-a2dp-loopback to start source without registry
      run e2e tests to confirm they still can detect packets
Change-Id: I2e64869eb6250bb788b73bae622eae54da5deac1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748944
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>