[docs] Adding more descriptive alt text to images
This adds a more descriptive alt text to images. This is done to improve
the utility of the documentation when used by the visually impared.

There is also an effort to preform more rigorous checks on markdown
documents which will disallow single word alt text values such as
"image" or "drawing".

Change-Id: I071f5c6a47bfafc059dee741d07c9996ee122217
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/741231
Reviewed-by: Kayce Basques <kayce@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>