[ui-input] Add _usecs to pointerinjector inspect values
Currently the values reported to inspect by the pointerinjector are
in nsecs, but that is not made clear anywhere. nsecs are also too
granular to really be helpful.
This CL changes the scale to usecs and adds a _usecs denominator to
the inspect field name.

Fixed: 113413
Change-Id: Ic0c2fa5ddafb98768e6c0475c73f5a82bf0af8b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749489
Reviewed-by: Sahil Madan <sahilmadan@google.com>
Commit-Queue: Sahil Madan <sahilmadan@google.com>