[fint] Improve no-op check error messages
Currently if the ninja no-op check fails, the error message that's
propagated to the build results page will say only "Ninja build did not
converge to no-op", and users must dig into the ninja explain stderr to
figure out which target is causing the problem.

The first line of the explain stderr is usually enough to explain the
failure, so we should include that in the error message.

This will also help with writing queries to analyze the rates of
different types of no-op failures.

Change-Id: I1632352068ce284f8360996950bd94149317b162
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748771
Reviewed-by: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>