[line_input] Don't flush input.
Previously when we changed terminal modes we did a flush. But the kind
of flush that was used discarded all buffered input in stdin. This
creates a race with user typing and has the effect of randomly
discarding input.

Change to only flushing the output on state changes. This is the
important part because that can get buffered and the output formatting
depends on the terminal mode being correct.

Change-Id: Ifdb0598a2f066ba00bf16b688d479e0921cc86c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749604
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>