[superproject] Roll llvm-project/libc [libc] add locale free strcoll
The strcoll function is intended to compare strings based on their
ordering in the current locale. Since the locale facilities have not yet
been added, a simple implementation that is the same as strcmp has been
added as a placeholder.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D136802

GitOrigin-RevId: a2037931c16af4ca695b6e718362acf4f99cb685
Original-Revision: f816f1a8df1f45d379e274631687a094247b7aad
Roller-URL: https://ci.chromium.org/b/8799024393589065121
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I2640d191dadb4abd303b2f0831bc2aa392b7e7dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749347