[inspector][gwp-asan] Ignore OOM errors.
GWP-ASan uses page faults to detect its own exceptions. However, if a
page fault occurs because the system ran out memory when trying to
access a gwp-asan tracked variable, the runtime will report it as a
gwp-asan error.

The error code for page faults can be found in the synth_code inside the
exception report's context.

This commit checks the error code of the page fault for out of memory
errors (`ZX_ERR_NO_MEMORY`), and if so, ignores the exception.

A test was added to simulate this code path. It is not possible to force
the system to generate an out-of-memory page fault via regular ways, so
instead we just change the synth_code of the exception report manually
for the test.

Fixed: 113375
Change-Id: I06c9e229ed85c91aaea92a47d30cfb2c62f83e7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748943
Reviewed-by: Francois Rousseau <frousseau@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>