[debugger] Allow thread state to be unknown.
The thread state and stack is tracked on the client and it can get
out-of-sync with the target. Previously, we cleared the stack when we
knew we were doing something that would invalidate it (pause or
continue), but the thread "state" was not correspondingly updated (since
we didn't actually know what really happened yet).

The result was occasional assertions about threads being stopped on
exceptions with no stacks. The rule is that if a thread is stopped on an
exception, the debug_agent will always supply a stack.

This converts the thread state to a std::optional so we can more clearly
denote that we don't know the current state and the empty stack will no
longer be an error.

The test in command_utils_unittest required that the objects be
registered with the console context to generate the error message. This
in turn required that the mock process have a target (previously it did
not). So MockProcess now takes a parent Target* and the tests that use
it were updated.

Change-Id: I8375a09aad155af9d03f599755e1b40f0010b170
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749725
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>