[audio][test] Async variant of TestServerAndClient
Split TestServerAndClient into sync and async variants.

The async variant is currently hardcoded to use natural types;
the sync variant is currently hardcoded to use wire types. We
will generalize this in a subsequent CL, so that this can be
specified with an additional template parameter.

Change-Id: Ia10bb5befb7462139e0ed1671652825d3cd2efba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742093
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>