[audio][test] Update test for new delay API
This updates audio_driver_tests for recent changes to
RingBuffer/GetProperties and the RingBufferProperties table,
and the addition of the WatchDelayInfo API.

The new WatchDelayInfo method uses the hanging-get pattern,
so the test expects a driver to always return a response
from the initial call. In the absence of additional
configuration that could change these delay values, it also
expects that a subsequent WatchDelayInfo call will be left
hanging.

Note that both external_delay and internal_delay are treated
as optional in this method.

RingBufferProperties.fifo_depth is now optional, since it
and RingBufferProperties.external_delay are now deprecated.
However, if either fifo_depth or external_delay values ARE
returned, then the driver must also provide matching
value(s) in its WatchDelayInfo response.

This CL also adds an explicit ring-buffer channel-close at
the end of RingBuffer-related test cases, along with a hard-
coded delay of 100ms. After closing the RingBuffer or
StreamConfig channels, various "real" audio drivers return
errors without this delay (and there is no other mechanism
that can be used to signal readiness).

Change-Id: Iafc517c5799946cb882a731c9a30e47c3c61ddb4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/746522
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>