[virtualization] Enable Rust virtio-net device
Compared this to the C++ device using iperf3 on Termina and a tethered
gLinux desktop. In both cases, throughput ranged from 400Mbps, to
750Mbps.

Both of these devices use a single thread to service both virtio queues,
so this is a reasonably respectable throughput.

Note that if you want to repro this, you need to patch in
fxrev.dev/749243 to work around fxbug.dev/113283.

Bug: 95485
Tested=iperf3

Change-Id: Ie7cb2d7a8fe512205c86fc4086491c0664c72245
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749244
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>