[gigaboot] Synthesise EFI memory attributes table.
On some systems, the EFI memory attributes table does not accurately
reflect all of the regions the EFI runtime services actually need.

For instance, on NUC11 the GetTime() call which we use to test EFI
relies on a MMIO region which is not included in the memory attributes
table.

This CL makes gigaboot synthesise an EFI memory attributes table by
combining the memory map returned by GetMemoryMap() and the memory
attributes table.

Bug: 113304
Test: k ut efi on NUC11.
Change-Id: I39b3d8810e6edbe5dea81299cf684cab7223650f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748288
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>