[build] Fix toolchain redirect
This group is intended to redirect test packages targets included in a
non-default toolchain to the corresponding target in the default
toolchain, but it was previously broken because it tried to use a
relative directory name rather than a relative target name.

I think this group will only ever be included when there's a bug in a GN
file that results in a target being included in the wrong toolchain.
That will now result in an `assert(is_fuchsia)` failing somewhere, which
is more helpful than a red herring "Unable to load <path>/BUILD.gn"
failure, which is what would happen before this fix.

Change-Id: Ia83aaf7f2cf9232eef9f9572b29695ca6cafbaba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750352
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>