[kernel][vm] Fix pin_count assert
When transitioning from AwaitingClean to Clean, we cannot assume that
pin_count will be zero. It is possible for a page to be pinned for read
after having been marked AwaitingClean, in which case it should be valid
to mark the page Clean while pinned. It is also possible for pages to be
in an intermediate pinned state as a write-pin will only attempt to
dirty pages after having pinned them, so a racing writeback should be
able to clean pages if the dirty request has not been resolved yet.

We can only assert that the pin count is zero when transitioning from
Dirty to AwaitingClean, since the writeback code explicitly forbids this
transition.

Fix the incorrect assert on Clean transition and add tests that would
have tripped it.

Test: core-pager-writeback
Bug: 106863
Bug: 63989
Run-All-Tests: true
Change-Id: Ie08600798571423ed914c88c95a968d226296ca1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749615
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>