[intel-i915-tgl] DP: Always train link on display initialization.
On Tiger Lake, now that we unconditionally reset and reinitialize
all the display devices, it is possible that a link is already
trained on display initialization. In that case, the driver should
not simply treat it as "trained" but should still do the initialization
procedure again.

Note that this CL also changes the behavior of Kaby Lake on display
re-initialization though we didn't see any side effects.

Test: HDMI port on Intel NUC11PAHi5
      Chromebook (Atlas) with bind rule changed

Bug: 109368
Change-Id: I19304977a8a6f7973ffa60c596ea2264d81bb0e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744873
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>