[virtualization] Added memory pressure events
Adding a way for the guest manager to react to the host memory pressure.
The guest manager will inflate the balloon to 90% of the free guest
memory when the host is under the memory pressure.
Balloon will be deflated once host's memory pressure gets back to
normal.

Bug: 100504

Change-Id: Ie1db2182e1c7953d5df77e52cc8d97033837dd26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731543
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Alexander Radchenko <alerad@google.com>