[bt][init] Use core-env for the #bluetooth-core realm
Currently, the bluetooth-core realm is not product variable. It is
defined in the base `core.cml` manifest.
`bt-rfcomm` and `bt-fastpair-provider` are product dependent children of
`bt-init`. To avoid churn during development, both packages are included
in the cached set of packages in eng builds. These two packages need the
full-resolver so that they can be re-resolved and updated ephemerally.

Use core-env for the BT realm so that the child components can be
resolved.

Bug: 108241
Test: manifest changes only, fx build, fx ota, verify
bt-fastpair-provider is resolved correctly.

Change-Id: Ie0769bd66b7c1072a565332d6b0e5a2677d80fed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750410
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>