[virtualization] Post error when VM fails to launch.
If the VM fails to launch, vsh will not report any feedback because the
error is not sent over FIDL. This updates termina_guest_manager to post
and error to the client in this situation.

We also add a new error for failing to allocate a disk for a guest since
this is a common failure mode.

Bug: 113232
Change-Id: Ife2f328c5f4a5e749521d97ffc44fece1bfb151d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748747
API-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Jonathan Bailey <jonathanbailey@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>