[fxfs] Fix cause of test timeout
futures::lock::Mutex is particularly unfair and it's easy to cause
starvation.  LockManager locks are also not guaranteed to be fair but
when locks are unlocked, *all* wakers are woken up and so who gets the
lock next is dependent on the executor, which is good enough to avoid
the starvation issue we were seeing with the test_unlink_with_open_race
test.

Fixed: 113620
Bug: 102659
Change-Id: I4374bd0ee38c74acbd73e2abef92e4d7d2d52dbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751082
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>