[fdf] Support DdkOpen and DdkClose in compat shim
Support the ability for DFv1 drivers to listen to when their associated
devfs node gets opened and closed when DFv2 is enabled and the driver is
running in a compat shim.

Properly close connections to vnode associated with instance devices
when the device is destructed in the DFv2 compat shim.

Don't modify DFv2 compat device's name when creating an outgoing name
for the device.

Bug: 102234
Bug: 112361

Change-Id: Ibe78b0aa62e5747235ca1e16658cbf9815e79d6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748783
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>