[forensics] Set signature for no memory page fault crashes
Programs page fault with a ZX_ERR_NO_MEMORY status when the system is
unable to allocate memory for them. Capture this reason in the crash
signature in order to prevent the signature from being based on the
stack trace, which is useless because the code is not fault for the
crash.

Bug: 113439

TESTED=Not able to b/c this code path requires a real exception to be
generated and we don't have a way to create a ZX_ERR_NO_MEMORY page
fault on demand

Change-Id: I36a7b3e3f4eedf2141d1b766d96de2fbb16ce34c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751342
Reviewed-by: Francois Rousseau <frousseau@google.com>
Commit-Queue: Alex Pankhurst <pankhurst@google.com>
Reviewed-by: Thomas Gales <tgales@google.com>