[xhci] Fix ERDP pointer read
This patch fixes an edge case where we might incorrectly read an event
TRB that HW hasn't handed control of over to us yet.

This occurs when we're reading a new segment for the first time, and so
validity checks should be done by completion code instead of cycle bit.
The buggy sequence is:
1. AdvanceErdp increments ERDP and returns a Control struct with its
   cycle bit set to !ccs_.
2. In HandleIRQ the cycle bit is checked and since it doesn't match ccs_
   we return.
3. We get a spurious interrupt.
4. We read from the ERDP, skipping the completion code check and not
   modifying the cycle bit. We enter the loop based on an invalid cycle
   bit check.

It's not clear whether this is benign (e.g. if we never get spurious
interrupts it's fine) but this change is at least more correct.

Bug: 107934
Change-Id: I3170ff0dcb40a2b8eaf0b2dc265deb9f49c617c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747289
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>