[packet] Simplify limited records representation
Simplify the representation of record formats which keep track of the
number of records which are expected to be parsed. Previously, an
entirely separate set of traits and utilities were provided to parse
these formats. In this commit, we represent the "normal" case - in which
no count is provided ahead of time - as a special case of the "general"
case - in which a count may be provided. This also allows us to remove
the `XxxBridge` types, which were a hack to avoid having multiple
blanket impls of the `RecordsImpl` trait. With the limited records
utilities removed, there can be a single blanket impl used by options
parsing.

Also previously, it was possible to encode a format in which it is an
error to provide more records than are expected using this count. It
turns out that no formats we support use this functionality, so in this
commit, it is removed.

Finally, remove some testing utilities which used `BufferView`s. These
utilities are not currently used, and should have been removed in
whatever previous commit removed their final use.

Change-Id: I0022f9e3b01dc9b47940c3d8eb2b11e499033d20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750804
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>