[netstack3] Update multicast membership w/ lock
When a datagram socket requests to join or leave a multicast group,
perform the operation while still holding the socket state lock. This
ensures that join and leave operations are atomic and moves us closer to
the goal of removing the lock protecting Ctx in the bindings.

Bug: 108008
Change-Id: Ib3d645a67cb08619dfe8ba2ea7008fe1c41a9a8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750500
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>