[docs][ffx] Removing manual ssh key generation steps.
These files are generated by `ffx` when flashing or running the emulator, and there is no longer a need for developers to generate the keys manually.

If there is a need by a specific developer to use specific keys, they will know how to obtain the keys they should be using, or how to generate them correctly for their use case.

Bug: 113911
Change-Id: I2c0efcb1c4f382ecbe18a454f651c0b1a4f0034b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752222
Reviewed-by: Wayne Piekarski <waynepie@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Kyo Lee <kyol@google.com>