[perfetto-bridge] Added an integration test
This test set up the topology

         test-harness
               |
               |
          trace_manager
           /        \
perfetto-bridge---test_producer

Then starts and stops tracing and checks if perfetto data exists in the
trace buffer.

It does not yet emit perfetto data itself which will come in a follow up
patch that implements perfetto::Producer.

It does check that it's able to correctly make a connection and that
it's able to see the data that perfetto emits on startup.

Change-Id: I8f5bbaaaedd8e483b55b43041b437658b622aecc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735074
Reviewed-by: Bryant Chandler <bryantchandler@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>