[emu] Add one-liner about network mode to 'start'
The attached bug describes how a user can be confused by the default
behavior of the --net flag in `ffx emu start`, leading to potential
support tickets. This CL adds some visibility around the resolution of
the "Auto" value by printing a single line to the terminal output after
resolution is complete. This line contains the mode selected, and a link
to a fuchsia.dev page with more information about networking modes. This
message is only printed when the "--net auto" flag is being resolved,
not when the networking mode is explicitly chosen by the user.

Bug: 113058
Change-Id: I1ca04ffe141a682a6c7795af32842a84e89fb013
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750372
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>