[netstack3] TCP cubic congestion control
This commit implements RFC 8312 for the TCP state machine in Core: it
implements the slow start and the congestion avoidance algorithm. Fast
convergence, fast retransmission and fast recovery are not in this
commit and will be addressed in a future commit. One difference worth
noting is that we count bytes instead of segments for our congestion
window as opposed to the RFC.

Bug: 111778
Change-Id: Iefd662f6036f1f844e0eb789ca17593352ca0ad8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747062
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>