[wlan][sme] Replace ScanTransaction with return from Scan
SME now returns a full scan result to the caller instead of a protocol
the caller must stream results from. The direct return is enabled by
using a new FIDL attribute to enable sending messages of unbounded
size. This new approach has a negligible impact on latency and is
obviously much more ergonomic. The latency of the returned results
is can almost be viewed as irrelevant because the scan itself requires
seconds to complete, and latency is on the order microseconds.

The ScanError type is removed. The additional message string in that
type was never used in a meaningful way, and its removal makes the
Scan() method more ergnomic for Rust.

This is an initial implementation of the FIDL large messages RFC
available at fxrev.dev/695625.

Test: Existing tests modified. Tests added for large scan results.
Bug: 88648, 107777

Change-Id: I0269360f64650bf97d1ce2b93ca7b2c7601fc44f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705311
Reviewed-by: Marc Khouri <mnck@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>