[bt][hfp] Fix crash formatting codecs as strings
Previously, if the AG recevied a codec ID from a peer that was not one
specified in the HFP spec (either 0x1 0x2), it would panic attempting to
format it as a string.  This means that noncompliant peers can crash
HFP. We are seeing a few of these a day in the field.

This change adds support for unknown codecs to the string formater for
code IDs and adds a test for this.

Test: fx test fuchsia-pkg://fuchsia.com/bt-hfp-audio-gateway-tests#meta/test-component.cm
Bug: fxb/113474
Change-Id: I276d00411f219b2947a9d2f0909c542b636204d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752983
Commit-Queue: Matthew Kehrt <kehrt@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>