[component][fuzzing] Refactor the spawn actions of ChildProcess
This CL reworks how the fdio_spawn_action_ts are created by ChildProcess
objects. These structs are passed to fdio_spawn, and allow callers to
add startup channels and specify how stdio should be set up.

The motivation for this change was to try to address a problem observed
with noisy fuzzers running within the test runner framework. When a
fuzzer generated excessive amounts of output, the test runner could
become stuck and stop forwarding that output to fuzz_ctl via the
fuzz-manager and test_manager. This change mitigates that outcome by
preventing stdout from being forwarded to the test runner.

Initially, this change also sought to exclude `fuchsia.logger.LogSink`
from the capabilities in the fuzzer sub-process's namespace. Eventually,
it was observed that excessive syslogs did *not* interfere with fuzzer
operation. The CL was modified to leave syslog intact and only change
how stdio is handled (albeit with a few more changes than might have
been included otherwise).

There are three critical details that are changed as a result of this
refactor that are worth noting:

 - ChildProcess's stdout and stderr threads now keep running even after
   the associated receiver has been closed in order to drain their
   pipes. They stop when the process stops.
 - ChildProcess now *does not* clone or transfer file descriptors by
   default. Callers of ChildProcess must explicitly configure that
   behavior, e.g. via AddStdinPipe() or CloneStderr().
 - LibFuzzerRunner explicitly adds pipes for stdin, stdout, and stderr,
   although it only uses stderr. This is to ensure the file descriptors
   are valid when libFuzzer re-executes itself and tries to clone all
   stdio.

Bug: 109100
Change-Id: I471c88572f26488dee54bb973eb6f2a78e549289
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750503
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>